+++ src/edu/mit/csail/sdg/alloy4graph/VizGraph.java
@@ -458,11 +458,11 @@ public final strictfp class VizGraph extends DiGraph {
         if (nodes.size()==0) { top=0; bottom=10; totalHeight=10; left=0; totalWidth=10; return; }
         if (fresh) { top=nodes.get(0).y()-nodes.get(0).getHeight()/2-5; bottom=nodes.get(0).y()+nodes.get(0).getHeight()/2+5; }
         // Find the leftmost and rightmost pixel
-        int minX = (int) (nodes.get(0).x() - nodes.get(0).getWidth()/2 - 5);
-        int maxX = (int) (nodes.get(0).x() + nodes.get(0).getWidth()/2 + nodes.get(0).getReserved() + 5);
+        int minX = nodes.get(0).x() - nodes.get(0).getWidth()/2 - 5;
+        int maxX = nodes.get(0).x() + nodes.get(0).getWidth()/2 + nodes.get(0).getReserved() + 5;
         for(VizNode n:nodes) {
-            int min = (int) (n.x() - n.getWidth()/2                   - 5); if (minX>min) minX=min;
-            int max = (int) (n.x() + n.getWidth()/2 + n.getReserved() + 5); if (maxX<max) maxX=max;
+            int min = n.x() - n.getWidth()/2                   - 5; if (minX>min) minX=min;
+            int max = n.x() + n.getWidth()/2 + n.getReserved() + 5; if (maxX<max) maxX=max;
         }
         for(VizEdge e:edges) if (e.getLabelW()>0 && e.getLabelH()>0) {
             int x1=e.getLabelX(), x2=x1+e.getLabelW()-1;
